env yml + src/output&setup/tests warns

+++ src/output/tests/writer-mpi.cpp
@@ -33,7 +33,11 @@ auto main(int argc, char* argv[]) -> int {
   try {
     using namespace ntt;
     auto writer = out::Writer("hdf5");
-    writer.defineMeshLayout({ size * 10 }, { rank * 10 }, { 10 }, false, Coord::Cart);
+    writer.defineMeshLayout({ static_cast<unsigned long>(size) * 10 },
+                            { static_cast<unsigned long>(rank) * 10 },
+                            { 10 },
+                            false,
+                            Coord::Cart);
     writer.defineFieldOutputs(SimEngine::SRPIC, { "E" });
 
     ndfield_t<Dim::_1D, 3> field { "fld", 10 + 2 * N_GHOSTS };
